requirements.txt
flask==2.3.3
transformers==4.53.0
torch==2.2.2
bleach==6.2.0
flask-limiter==2.9.2
flask-talisman==1.1.0
gunicorn==23.0.0
requests==2.32.5
huggingface-hub>=0.22.0
accelerate>=0.29.0
python-dotenv>=1.0.0
